Apply refactoring patch
