+++ hudi-client/hudi-spark-client/src/main/java/org/apache/hudi/io/storage/HoodieSparkLanceWriter.java
@@ -18,14 +18,13 @@
 
 package org.apache.hudi.io.storage;
 
-import org.apache.hudi.SparkAdapterSupport$;
-import org.apache.hudi.avro.HoodieBloomFilterWriteSupport;
 import org.apache.hudi.common.bloom.BloomFilter;
 import org.apache.hudi.common.engine.TaskContextSupplier;
 import org.apache.hudi.common.model.HoodieKey;
 import org.apache.hudi.common.model.HoodieRecord;
 import org.apache.hudi.common.util.Option;
 import org.apache.hudi.io.lance.HoodieBaseLanceWriter;
+import org.apache.hudi.io.storage.row.HoodieBloomFilterRowWriteSupport;
 import org.apache.hudi.io.storage.row.HoodieInternalRowFileWriter;
 import org.apache.hudi.storage.HoodieStorage;
 import org.apache.hudi.storage.StoragePath;
@@ -34,7 +33,6 @@ import com.lancedb.lance.spark.arrow.LanceArrowWriter;
 import lombok.AllArgsConstructor;
 import org.apache.arrow.vector.VectorSchemaRoot;
 import org.apache.arrow.vector.types.pojo.Schema;
-import org.apache.spark.sql.HoodieUTF8StringFactory;
 import org.apache.spark.sql.catalyst.InternalRow;
 import org.apache.spark.sql.types.StructType;
 import org.apache.spark.sql.util.LanceArrowUtils;
@@ -69,7 +67,7 @@ public class HoodieSparkLanceWriter extends HoodieBaseLanceWriter<InternalRow>
   private final UTF8String instantTime;
   private final boolean populateMetaFields;
   private final Function<Long, String> seqIdGenerator;
-  private final Option<HoodieBloomFilterLanceWriteSupport> bloomFilterWriteSupportOpt;
+  private final Option<HoodieBloomFilterRowWriteSupport> bloomFilterRowWriteSupportOpt;
 
   /**
    * Constructor for Spark Lance writer.
@@ -90,7 +88,7 @@ public class HoodieSparkLanceWriter extends HoodieBaseLanceWriter<InternalRow>
                                 HoodieStorage storage,
                                 boolean populateMetaFields,
                                 Option<BloomFilter> bloomFilterOpt) throws IOException {
-    super(file, DEFAULT_BATCH_SIZE, bloomFilterOpt.map(HoodieBloomFilterLanceWriteSupport::new));
+    super(file, DEFAULT_BATCH_SIZE, bloomFilterOpt.map(HoodieBloomFilterRowWriteSupport::new));
     this.sparkSchema = sparkSchema;
     this.arrowSchema = LanceArrowUtils.toArrowSchema(sparkSchema, DEFAULT_TIMEZONE, true, false);
     this.fileName = UTF8String.fromString(file.getName());
@@ -100,12 +98,8 @@ public class HoodieSparkLanceWriter extends HoodieBaseLanceWriter<InternalRow>
       Integer partitionId = taskContextSupplier.getPartitionIdSupplier().get();
       return HoodieRecord.generateSequenceId(instantTime, partitionId, recordIndex);
     };
-    if (super.bloomFilterWriteSupportOpt.isPresent()) {
-      this.bloomFilterWriteSupportOpt = Option
-          .of((HoodieBloomFilterLanceWriteSupport) super.bloomFilterWriteSupportOpt.get());
-    } else {
-      this.bloomFilterWriteSupportOpt = Option.empty();
-    }
+    this.bloomFilterRowWriteSupportOpt = super.bloomFilterWriteSupportOpt
+        .map(s -> (HoodieBloomFilterRowWriteSupport) s);
   }
 
   /**
@@ -148,10 +142,10 @@ public class HoodieSparkLanceWriter extends HoodieBaseLanceWriter<InternalRow>
   public void writeRowWithMetadata(HoodieKey key, InternalRow row) throws IOException {
     if (populateMetaFields) {
       UTF8String recordKey = UTF8String.fromString(key.getRecordKey());
-      bloomFilterWriteSupportOpt.ifPresent(bloomFilterWriteSupport ->
-              bloomFilterWriteSupport.addKey(recordKey));
+      bloomFilterRowWriteSupportOpt.ifPresent(bloomFilterWriteSupport ->
+          bloomFilterWriteSupport.addKey(recordKey));
       updateRecordMetadata(row, recordKey, key.getPartitionPath(), getWrittenRecordCount());
-      super.write(row);
+      super.write(row.copy());
     } else {
       super.write(row);
     }
@@ -159,14 +153,14 @@ public class HoodieSparkLanceWriter extends HoodieBaseLanceWriter<InternalRow>
 
   @Override
   public void writeRow(String recordKey, InternalRow row) throws IOException {
-    bloomFilterWriteSupportOpt.ifPresent(bloomFilterWriteSupport ->
+    bloomFilterRowWriteSupportOpt.ifPresent(bloomFilterWriteSupport ->
             bloomFilterWriteSupport.addKey(UTF8String.fromString(recordKey)));
     super.write(row.copy());
   }
   
   @Override
   public void writeRow(UTF8String key, InternalRow row) throws IOException {
-    bloomFilterWriteSupportOpt.ifPresent(bloomFilterWriteSupport ->
+    bloomFilterRowWriteSupportOpt.ifPresent(bloomFilterWriteSupport ->
             bloomFilterWriteSupport.addKey(key));
     super.write(row.copy());
   }
@@ -233,35 +227,6 @@ public class HoodieSparkLanceWriter extends HoodieBaseLanceWriter<InternalRow>
     @Override
     public void finishBatch() {
       lanceArrowWriter.finish();
-    }
-  }
-
-  /**
-   * Bloom filter write support implementation for Lance file format.
-   * Handles UTF8String record keys specific to Spark InternalRow.
-   */
-  private static class HoodieBloomFilterLanceWriteSupport extends HoodieBloomFilterWriteSupport<UTF8String> {
-
-    private static final HoodieUTF8StringFactory UTF8STRING_FACTORY =
-            SparkAdapterSupport$.MODULE$.sparkAdapter().getUTF8StringFactory();
-
-    public HoodieBloomFilterLanceWriteSupport(BloomFilter bloomFilter) {
-      super(bloomFilter);
-    }
-
-    @Override
-    protected byte[] getUTF8Bytes(UTF8String key) {
-      return key.getBytes();
-    }
-
-    @Override
-    protected UTF8String dereference(UTF8String key) {
-      return key.clone();
-    }
-
-    @Override
-    protected int compareRecordKey(UTF8String a, UTF8String b) {
-      return UTF8STRING_FACTORY.wrapUTF8String(a).compareTo(UTF8STRING_FACTORY.wrapUTF8String(b));
     }
   }
 }
